Rename 'bun up' to 'bun db' (#141)
### Description

<!-- Provide a comprehensive description here about what your PR aims to
solve. -->

<!-- You may also add additional context -->

Update the "bun up" command to "bun db" since it's used to spin up a db
container. It's a better name. If we use "bun docs" to spin up a docs
server, we should use "bun db" to spin up a db server.

---

### PR Checklist

<!-- Please do not remove this section -->

<!-- Mark each item with an "x" ([ ] becomes [x]) -->

- [x] Read the Developer's Guide in
[CONTRIBUTING.md](https://github.com/agnyz/bedstack/blob/main/CONTRIBUTING.md)
- [x] Use a concise title to represent the changes introduced in this PR
- [x] Provide a detailed description of the changes introduced in this
PR, and, if necessary, some screenshots
- [x] Reference an issue or discussion where the feature or changes have
been previously discussed
- [x] Add a failing test that passes with the changes introduced in this
PR, or explain why it's not feasible
- [x] Add documentation for the feature or changes introduced in this PR
to the docs; you can run them with `bun docs`


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Documentation**
- Updated setup instructions to now use the `bun db` command for
initiating the database service, replacing the previous `bun up` command
in both the quickstart and getting-started guides.
- **Chores**
- Adjusted the underlying command reference in the configuration to
align with the new `bun db` command.
- **Bug Fixes**
- Removed the `app` service configuration from the Docker Compose setup.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>